Added Windows support to HLSCore interface  - Moved operating system specific defines to Defines.h  - Added casts to ensure compatibility with 32-bit compilers  - Added warning that interrupts are not implemtend for Windows

+++ API/include/IP_Cores/HLSCore.h
@@ -161,7 +161,7 @@ public:
 	void SetDataAddr(const uint64_t& offset, const Memory& mem, const AddressType& addrType = AddressType::BIT_64)
 	{
 		if (addrType == AddressType::BIT_32)
-			setDataAddr<uint32_t>(offset, mem.GetBaseAddr());
+			setDataAddr<uint32_t>(offset, static_cast<uint32_t>(mem.GetBaseAddr()));
 		else
 			setDataAddr<uint64_t>(offset, mem.GetBaseAddr());
 	}
